feat(gfx): breaking change - need root signature when binding descriptor tables
Breaking Change: We could cache this during descriptor table create time but tables/sets can be used with any root sig, I want to keep things as decoupled as possible and we will soon switch to using one single master compute/gfx root sig for entire engine so this doesn't matter, renderer will own that. For now the API is as the backend needs it, no hidden caching. Let the frontend abstract the usage/creation of root sigs